requirements.txt
flask
flask-session
flask-mysqldb
together
